Copy Doxygen configuration file to doc/
Autotools m4 scripts are using the following Doxygen template file:
`doxygen_libdvbv5.cfg`. In order to support Doxygen documentation in
meson a different template file is needed, as the syntax for variable
substitution is not shared between the two build systems. While m4 uses
$(variable), meson supports ${variable} and @variable@, see [0].

Let's create a copy of the Doxygen documentation configuration file
`doxygen_libdvbv5.cfg`, so it can be customized for meson.

[0] https://mesonbuild.com/Reference-manual.html#configure_file

Signed-off-by: Ariel D'Alessandro <ariel.dalessandro@collabora.com>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>